move ingest endpoint to internal routes
